HW1: Improve lab performance
Dynamips and Dynagen by default use memory mapped files to provide
RAM for the emulated routers. That might allow to provide more
virtual RAM than physical RAM is available on the host. But with
usual RAM sizes this is not necessary (and the virtual memory of
the host should provide the same service via swap space anyway).

A side effect of using memory mapped files is that the file contents
are actually written to disk, even if this is not necessary. That
results in useless I/O activity badly affecting system performance.
This is especially true for Linux with its well known problems under
moderate to high I/O load.

Disabling Dynamips' use of memory mapped files for providing router
RAM results in a much more responsive host system while the lab is
running.